refactor: re-export all features
