+++ apps/web/app/[slug]/_features/index.ts
@@ -1,0 +1,1 @@
+export { Content } from './Content'
