+++ apps/web/app/_components/index.ts
@@ -1,4 +1,8 @@
-export * from './Icons'
-export * from './MarkdownContent'
-export * from './MyAvatar'
-export * from './SocialLink'
+/** @public */
+export { MoonIcon, SunIcon } from './Icons'
+/** @public */
+export { MarkdownContent } from './MarkdownContent'
+/** @public */
+export { MyAvatar } from './MyAvatar'
+/** @public */
+export { SocialLink } from './SocialLink'
