+++ apps/web/app/_features/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './Header'
+export { ColorModeScript } from './ColorModeScript'
