+++ apps/web/app/_utils/index.ts
@@ -1,3 +1,4 @@
-export * from './constants'
-export * from './getSlugs'
-export * from './rootJoin'
+/** @public */
+export { env, externalLinks } from './constants'
+/** @public */
+export { rootJoin } from './rootJoin'
