+++ apps/web/app/_utils/server.ts
@@ -1,0 +1,2 @@
+/** @public */
+export { getSlugs } from './getSlugs'
