+++ apps/web/app/articles/[slug]/_features/index.ts
@@ -1,0 +1,1 @@
+export { Article } from './Article'
