+++ apps/web/app/articles/_features/index.ts
@@ -1,0 +1,2 @@
+export { ArticleList } from './ArticleList'
+export { Footer } from './Footer'
