+++ apps/web/app/articles/page.tsx
@@ -1,4 +1,4 @@
-import { ArticleList } from './_features/ArticleList'
+import { ArticleList } from './_features'
 
 export default function Page() {
   return <ArticleList />
